requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==6.2.1
lxml==4.9.2
outcome==1.2.0
parse==1.19.0
pycodestyle==2.10.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
requests-html==0.10.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
zipp==3.15.0
